+++ CHANGES.rst
@@ -29,6 +29,7 @@ Bugfixes
 - Fix error in HTTP API when Basic auth headers are present (:issue:`4123`,
   thanks :user:`uxmaster`)
 - Fix incorrect font size in some room booking dropdowns (:issue:`4156`)
+- Add missing email validation in some places (:issue:`4158`)
 
 Version 2.2.4
 -------------